# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: MWRD B2B Procurement Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rfqs:
    get:
      operationId: listRFQs
      tags: [rfq]
      summary: List RFQs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RFQListResponse"
    post:
      operationId: createRFQ
      tags: [rfq]
      summary: Create a custom-request RFQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRFQBody"
      responses:
        "201":
          description: RFQ created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RFQ"
  /rfqs/{id}:
    get:
      operationId: getRFQ
      tags: [rfq]
      summary: Get a single RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: RFQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RFQ"
        "404":
          description: Not found
  /rfqs/{id}/quotes:
    get:
      operationId: listQuotesForRFQ
      tags: [rfq]
      summary: List quotes for an RFQ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
  /rfqs/{id}/award/full-basket:
    post:
      operationId: awardFullBasket
      tags: [rfq]
      summary: Award entire RFQ to one supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwardFullBasketBody"
      responses:
        "201":
          description: CPO + SPO created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/POPairResponse"
  /rfqs/{id}/award/per-line:
    post:
      operationId: awardPerLine
      tags: [rfq]
      summary: Award individual lines to different suppliers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwardPerLineBody"
      responses:
        "201":
          description: PO pairs created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/POPairResponse"
  /quotes:
    get:
      operationId: listMyQuotes
      tags: [quote]
      summary: List supplier's own quotes
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteListResponse"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quote]
      summary: Get a single quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
    patch:
      operationId: editQuote
      tags: [quote]
      summary: Edit quote before sending
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditQuoteBody"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}/send:
    post:
      operationId: sendQuote
      tags: [quote]
      summary: Send quote to client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quote submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /orders:
    get:
      operationId: listOrders
      tags: [order]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of POs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PO"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [order]
      summary: Get order detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PO"
        "404":
          description: Not found
...
  /orders/{id}/approve:
    post:
      operationId: approveOrder
      tags: [order]
      summary: Approve a pending order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveOrderBody"
      responses:
        "200":
          description: Approval task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalTask"
  /orders/{id}/reject:
    post:
      operationId: rejectOrder
      tags: [order]
      summary: Reject a pending order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectOrderBody"
      responses:
        "200":
          description: Approval task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalTask"
...
  /orders/{id}/grn:
    post:
      operationId: createGRN
      tags: [order]
      summary: Create goods receipt note (client only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGRNBody"
      responses:
        "201":
          description: GRN created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GRN"
...
components:
  schemas:
...
    RFQItem:
      type: object
      properties:
        id:
          type: string
        rfq_id:
          type: string
        master_product_id:
          type: string
          nullable: true
        free_text_name:
          type: string
          nullable: true
        description:
          type: string
        qty:
          type: integer
        unit:
          type: string
        pack_type:
          type: string
          nullable: true
      required: [id, description, qty, unit]
    RFQ:
      type: object
      properties:
        id:
          type: string
        rfq_number:
          type: string
        client_company_id:
          type: string
        title:
          type: string
        description:
          type: string
        delivery_city:
          type: string
        delivery_date:
          type: string
        status:
          type: string
        source:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/RFQItem"
        created_at:
          type: string
        expires_at:
          type: string
      required: [id, rfq_number, title, status]
    RFQListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RFQ"
        total:
          type: integer
      required: [data, total]
    CreateRFQBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        delivery_city:
          type: string
        delivery_date:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              master_product_id:
                type: string
                nullable: true
              free_text_name:
                type: string
                nullable: true
              description:
                type: string
              qty:
                type: integer
              unit:
                type: string
              pack_type:
                type: string
                nullable: true
            required: [description, qty, unit]
      required: [title, delivery_city, delivery_date, items]
    QuoteItem:
      type: object
      properties:
        id:
          type: string
        quote_id:
          type: string
        rfq_item_id:
          type: string
        supplier_unit_price_sar:
          type: number
        final_unit_price_sar:
          type: number
        qty_available:
          type: integer
        lead_time_days:
          type: integer
        notes:
          type: string
        declined:
          type: boolean
      required: [id, rfq_item_id, final_unit_price_sar, qty_available]
    Quote:
      type: object
      properties:
        id:
          type: string
        quote_number:
          type: string
        rfq_id:
          type: string
        supplier_company_id:
          type: string
        status:
          type: string
        is_auto_generated:
          type: boolean
        valid_until:
          type: string
        lead_time_days:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        notes:
          type: string
        submitted_at:
          type: string
          nullable: true
        admin_held:
          type: boolean
      required: [id, quote_number, rfq_id, status]
    QuoteListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Quote"
        total:
          type: integer
      required: [data, total]
    EditQuoteBody:
      type: object
      properties:
        notes:
          type: string
        lead_time_days:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              supplier_unit_price_sar:
                type: number
              qty_available:
                type: integer
              notes:
                type: string
              declined:
                type: boolean
    AwardFullBasketBody:
      type: object
      properties:
        quote_id:
          type: string
      required: [quote_id]
    LineSelection:
      type: object
      properties:
        quote_id:
          type: string
        quote_item_id:
          type: string
      required: [quote_id, quote_item_id]
    AwardPerLineBody:
      type: object
      properties:
        selections:
          type: array
          items:
            $ref: "#/components/schemas/LineSelection"
      required: [selections]
    POItem:
      type: object
      properties:
        id:
          type: string
        master_product_id:
          type: string
        name_en:
          type: string
        qty:
          type: integer
        pack_type:
          type: string
        unit_price_sar:
          type: number
        total_sar:
          type: number
      required: [id, name_en, qty, unit_price_sar, total_sar]
    PO:
      type: object
      properties:
        id:
          type: string
        po_number:
          type: string
        type:
          type: string
        transaction_ref:
          type: string
        client_company_id:
          type: string
        supplier_company_id:
          type: string
        status:
          type: string
        total_sar:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/POItem"
        created_at:
          type: string
      required: [id, po_number, type, status, total_sar]
    POPairResponse:
      type: object
      properties:
        cpo:
          $ref: "#/components/schemas/PO"
        spo:
          $ref: "#/components/schemas/PO"
      required: [cpo, spo]
    ApprovalTask:
      type: object
      properties:
        id:
          type: string
        po_id:
          type: string
        approver_user_id:
          type: string
        status:
          type: string
        order_in_chain:
          type: integer
        decided_at:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
      required: [id, po_id, status, order_in_chain]
...
    ApproveOrderBody:
      type: object
      properties:
        task_id:
          type: string
        note:
          type: string
          default: ""
      required: [task_id]
    RejectOrderBody:
      type: object
      properties:
        task_id:
          type: string
        note:
          type: string
      required: [task_id, note]
...
    GRNItem:
      type: object
      properties:
        id:
          type: string
        master_product_id:
          type: string
        name_en:
          type: string
        qty_received:
          type: integer
        condition:
          type: string
      required: [name_en, qty_received, condition]
    GRN:
      type: object
      properties:
        id:
          type: string
        grn_number:
          type: string
        cpo_id:
          type: string
        dn_id:
          type: string
        received_at:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/GRNItem"
        notes:
          type: string
      required: [id, grn_number, cpo_id, dn_id, received_at]
    CreateGRNBody:
      type: object
      properties:
        dn_id:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              master_product_id:
                type: string
              name_en:
                type: string
              qty_received:
                type: integer
              condition:
                type: string
                enum: [ok, damaged, short]
            required: [name_en, qty_received, condition]
      required: [dn_id, items]
...