# lib/api-zod/src/generated/types/pOItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MWRD B2B Procurement Platform API
 * OpenAPI spec version: 0.2.0
 */

export interface POItem {
  id: string;
  master_product_id?: string;
  name_en: string;
  qty: number;
  pack_type?: string;
  unit_price_sar: number;
  total_sar: number;
}